renamed mismatching file names for statistics tests
